Remove dtype=np.float32 annotations, why not use float64?!

+++ pylat/calccomf.py
@@ -4,15 +4,15 @@ from numba import njit
 
 @njit
 def calccom(n, nummol, x, y, z, mol, amass, molmass, Lx, Ly, Lz, Lx2, Ly2, Lz2):
-    comxt = np.zeros(nummol, dtype=np.float32)
-    comyt = np.zeros(nummol, dtype=np.float32)
-    comzt = np.zeros(nummol, dtype=np.float32)
-    xt = np.zeros(nummol, dtype=np.float32)
-    yt = np.zeros(nummol, dtype=np.float32)
-    zt = np.zeros(nummol, dtype=np.float32)
-    ux = np.zeros(n, dtype=np.float32)
-    uy = np.zeros(n, dtype=np.float32)
-    uz = np.zeros(n, dtype=np.float32)
+    comxt = np.zeros(nummol)
+    comyt = np.zeros(nummol)
+    comzt = np.zeros(nummol)
+    xt = np.zeros(nummol)
+    yt = np.zeros(nummol)
+    zt = np.zeros(nummol)
+    ux = np.zeros(n)
+    uy = np.zeros(n)
+    uz = np.zeros(n)
     for i in range(n):
         mol_i = int(mol[i] - 1)
         if xt[mol_i] == 0.0:
